+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 - Added method to get all the selected facets - instance.getAllSelectedFacets()
 - Added a method to get the total count of selected facets - instance.getSelectedFacetsCount()
 - Added a method to get the total count of selected facet values - instance.getSelectedFacetValueCount()
-- Added wrapperTag options to products config
+- Added tagName, events and htmlAttributes to each config so that wrapper element created by sdk can be easily customised by customer
+- Added browseQueryParam to set the default category param for the category/ browse calls.
 
 
 ### Changed